Add basic test for audio generation
