+++ package.json
@@ -7,7 +7,7 @@
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
-    "test": "echo \"No tests defined\" && exit 0"
+    "test": "node --test"
   },
   "dependencies": {
     "react": "19.1.0",
